pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.2",
    "langchain-groq>=0.3.8",
    "langchain>=0.3.27",
    "langgraph>=0.6.7",
    "openai>=1.108.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sendgrid>=6.12.4",
    "streamlit>=1.49.1",
    "streamlit-chat>=0.1.1",
    "uvicorn>=0.35.0",
    "groq>=0.31.1",
    "email-validator>=2.3.0",
]
